Refactor text editor to React component
