+++ public/index.html
@@ -5,6 +5,8 @@
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.12.4/jquery.min.js"></script>
     <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"></script>
+    <script src="//cdn.quilljs.com/1.0.6/quill.min.js"></script>
+    <script src="https://cdn.socket.io/socket.io-1.4.5.js"></script>
     <link rel="stylesheet" href="public/style.css">
     <link href="//cdn.quilljs.com/1.0.6/quill.snow.css" rel="stylesheet">
   </head>
@@ -14,9 +16,6 @@
     <script src="public/bundle.js"></script>
     <script src="https://code.jquery.com/jquery-3.1.0.min.js" integrity="sha256-cCueBR6CsyA4/9szpPfrX3s49M9vUU5BgtiJj06wt/s=" crossorigin="anonymous"></script>
     <script src='public/adapter.js'></script>    
-    <script src="https://cdn.socket.io/socket.io-1.4.5.js"></script>
     <script src='public/videocall.js'></script>
-    <script src="//cdn.quilljs.com/1.0.6/quill.min.js"></script>
-    <script src="public/quill.js"></script>
   </body>
 </html>
